requirements.txt
python-vlc
SpeechRecognition
wikipedia
pafy
pyttsx3
pyautogui
selenium
playsound
pillow
gtts
youtube-dl
libvlc.dll
pygame
pyttsx3
ibm-watson>=5.3.0
##FLAC for os